fix(orchestrator): resolve RNG desync and float precision drift in parity tests

+++ src/aether/representation.py
@@ -29,7 +29,7 @@ class Representation:
     encoder_id: str
 
     def __post_init__(self) -> None:
-        vector = np.asarray(self.vector, dtype=np.float32)
+        vector = np.asarray(self.vector, dtype=np.float64)
 
         if vector.ndim != 1:
             raise ValueError("Representation vector must be 1-D")
